Turns out that SVN doesn't do anything right.
I initially thought leaving in the default line :

---This line, and those below, will be ignored--

was fine because it appeared that SVN did the smart thing and respected
that statement.  Unfortunately, it can't handle the commit file if it's
got line endings that aren't homogenous.. FML.  Instead of running
dos2unix or something I'm just reverting to doing the removal myself.